view exercises. do not allow submissions or viewing the forum if not authenticated
